Mixing on torus (Houdini)
The user can set up the slope for the line in the 2-dimensional real space. After the line is created, the curve on the torus based on the line is also created. This file provides the animation of the growing curve from the starting point to the end (which is the starting point again).